Trigger Tewi to remind user that their waifu loves them
A classic lainbot feature returns! This was fun and cute.

Interestingly enough, this is the first actual regex in tewibot!
Despite how much Lain used to rely on them, it's actually been
easier for me to use Golang's built-in strings library. Just
goes to show how the language you use affects how you approach the
problem.