Send tags in URL rather than request body
For some reason, they changed the way the API works - you can't send
parameters in the body any more for GET requests. Thankfully, they do
work if you send them in the URL itself *shrug*

This is what I get for relying on a bunch of anime perverts for my
anime pervert program.